KnowledgeServer/requirements.txt
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
filelock==3.17.0
Flask==3.1.0
fsspec==2024.9.0
gunicorn==23.0.0
#huggingface-hub==0.29.3
idna==3.10
importlib-metadata>=3.6; python_version < "3.10"
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
MarkupSafe==3.0.2
mpmath==1.3.0
networkx>=3
packaging==24.2
pandas==2.2.3
pillow==11.1.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
scikit-learn==1.6.1
scipy>1
sentence-transformers==3.4.1
six==1.17.0
sqlean.py==3.47.0
sqlite-vec==0.1.6
sympy==1.13.3
threadpoolctl==3.5.0
tokenizers==0.21.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
Werkzeug==3.1.3
